HOTT-1361 Move the MeasureConditionPresenter
Straight move down to the model layer so it can used from the Permutations Calculator
